NWGHS Robotics Update 0.0.1
Added main page
Added a nav bar
Added a search bar, and dropdown (Not functional yet)
Added a documentation page
Added pages for game plan, arm, main chassis, software, drone launcher, and outreach.
All pages have a temporary image of one of the team captains on our team. This image can be replaced with an image of the said webpage later on.